requirements.txt
# general modules
django==5.2.12
python-decouple
pillow
djangorestframework

# third party modules
markdown
django-filter
drf-yasg[validation]
djangorestframework-simplejwt 
drf-spectacular
psycopg[binary]==3.2.9
asgiref==3.9.1


# deployment module
gunicorn>=21,<23

# email third party module
django-mail-templated

# refactoring & test
black
flake8
pytest
pytest_django
pytest-cov
